refactor: address review feedback for scope cleanup
- Replace `as MemberRole` type assertions with `memberRoleSchema.parse()`
  for runtime validation in scope-member-service and resolve-scope
- Rename `org-test-helpers.ts` to `clerk-org-mock.ts` for consistency
- Rename contract key `scopeMemberContract.status` to `getMembers`
- Rename `orgScope` variable to `teamScope` in invite test
